Brooming (#17)
* Cleaning Commands

* Making Folder

* Cleaning Imports and Commenting
