+++ src/main/java/frc/robot/commands/OneBar/OnebarDown.java
@@ -2,12 +2,10 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.OneBar;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.OnebarSubsystem;
-import edu.wpi.first.wpilibj.DutyCycleEncoder;
-import frc.robot.Constants;
 
 public class OnebarDown extends CommandBase {
   private final OnebarSubsystem onebarsubsystem;
