+++ src/main/java/frc/robot/commands/OneBar/OnebarIn.java
@@ -2,15 +2,16 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
-import frc.robot.subsystems.OnebarSubsystem;
+package frc.robot.commands.OneBar;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
+import frc.robot.subsystems.OnebarSubsystem;
 
-public class ArmRotStop extends CommandBase {
+public class OnebarIn extends CommandBase {
+  /** Creates a new OnebarIn. */
   private final OnebarSubsystem onebarsubsystem;
-  /** Creates a new ArmRotStop. */
-  public ArmRotStop(OnebarSubsystem subsystem) {
+
+  public OnebarIn(OnebarSubsystem subsystem) {
     onebarsubsystem = subsystem;
     // Use addRequirements() here to declare subsystem dependencies.
     addRequirements(subsystem);
@@ -18,12 +19,13 @@ public class ArmRotStop extends CommandBase {
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() {}
+  public void initialize() {
+  }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    onebarsubsystem.armRotationStop();
+    onebarsubsystem.armIn();
   }
 
   // Called once the command ends or is interrupted.
